trouble with beans - this works first

+++ src/main/java/com/bae/business/service/ActivityService.java
@@ -1,20 +1,20 @@
 package com.bae.business.service;
 
 public interface ActivityService {
-	
+
 	// Create
-	String createActivity(String userEmail,String activityLog);
+	String createActivity(String userEmail, String activityLog);
 
 	// Read
-	String getAnActivity(String userEmail,Long id);
+	String getAnActivity(String userEmail, int id);
 
 	String getAllActivities(String userEmail);
 
-	String getAllActivitiesByCategory(String userEmail,String category);
+	String getAllActivitiesByCategory(String userEmail, String category);
 
 	// Update
-	String updateActivity(String userEmail,String activityLog, Long id);
+	String updateActivity(String userEmail, String activityLog, int id);
 
 	// Delete
-	String deleteActivity(String userEmail,Long id);
+	String deleteActivity(String userEmail, int id);
 }
